src/@core/infra/db/typeorm/route.schema.spec.ts
import { title } from 'process';
import { DataSource } from 'typeorm'
import { Route } from '../../../domain/route.entity'
import { RouteSchema } from './route.schema';

const dataSource = new DataSource({
  type: 'sqlite',
  database: ':memory:',
  synchronize: true,
  logging: false,
  entities: [RouteSchema],
})

beforeEach(async () => {
  await dataSource.initialize();
});
afterEach(async () => {
  await dataSource.destroy();
});

describe('RouteSchema Tests', () => {
  test('should create a new route', async () => {
    const route = Route.create({
      title: 'Minha primeira rota',
      startPosition: { lat: -23.5489, lng: -46.6388 },
      endPosition: { lat: -23.55, lng: -46.64 },
      points: [
        { lat: -23.5489, lng: -46.6388 }
      ],
    })
    const routeRepo = dataSource.getRepository(Route)
    await routeRepo.save(route)
    console.log(await routeRepo.findBy( {title: 'Minha primeira rota'} ))
    expect(route.id).toBeDefined();
    expect(route.title).toBe('Minha primeira rota');
    expect(route.startPosition).toStrictEqual({ lat: -23.5489, lng: -46.6388 });
    expect(route.endPosition).toStrictEqual({ lat: -23.55, lng: -46.64 });
    expect(route.points).toStrictEqual([
      { lat: -23.5489, lng: -46.6388 }
    ]);
    expect(await routeRepo.findOneBy({ id: route.id })).toStrictEqual(route);
  })
});